Time: 1 ms (46.11%) | Memory: 63 MB (48.02%) - LeetSync

+++ 11-container-with-most-water/container-with-most-water.cpp
@@ -18,6 +18,6 @@ public:
             }
 
         }
-        return maxArea;
+        return (int)maxArea;
     }
 };
